george_compiler v0.1.0 GeorgeCompiler.SMC.Attribution

Functions

attrib(_, s, m, c)

Aplicação da atribuição retirando de S o valor e o nome da variavel para colocá-los no mapa M

C := E < m v S, M, := C > ⇒ < S, M [m/v], C >

attribution_decompose_tree(tree, s, m, c)

Decomposição da aŕvore de atribuição.

C := I < S, M, v := e C > ⇒ < v S, M, e := C >

get_variable_value(var, m)

Recupera o valor de uma variavel dado um mapa M

is_attribution(operation)

Retorna true caso a operação seja attrib.